requirements.txt
itchat_uos==1.5.0.dev0
openai==0.27.4
pandas==1.1.5
xlrd==1.2.0